requirements.txt
tensorflow-cpu
fastapi
Pillow
starlette
gunicorn
uvicorn
numpy
python-multipart
firebase_admin
google.cloud
uuid